Separated pages and navigation implemented, not yet with layout component
