+++ components/Layout/Layout.jsx
@@ -1,0 +1,10 @@
+import Navigation from "../Navigation/Navigation";
+import React from "react";
+
+export default function Layout() {
+  return (
+    <nav>
+      <Navigation />
+    </nav>
+  );
+}
