+++ pages/_app.js
@@ -1,10 +1,32 @@
 import GlobalStyle from "../styles";
+import useSWR from "swr";
+import Layout from "@/components/Layout/Layout";
+
+const URL = "https://example-apis.vercel.app/api/art";
+
+const fetcher = async (url) => {
+  const res = await fetch(url);
+
+  if (!res.ok) {
+    const error = new Error("An error occurred while fetching the data.");
+    error.info = await res.json();
+    error.status = res.status;
+    throw error;
+  }
+  return res.json();
+};
 
 export default function App({ Component, pageProps }) {
+  const { data, error, isLoading } = useSWR(URL, fetcher);
+  console.log("data: ", data);
+
+  if (error) return <div>failed to load</div>;
+  if (isLoading) return <div>loading...</div>;
+
   return (
     <>
       <GlobalStyle />
-      <Component {...pageProps} />
+      <Component {...pageProps} piecesData={data} />
     </>
   );
 }
